simplify resources
